# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Gin personal blog backend API.",
        "title": "MyBlog API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/articles/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员创建文章，文章初始状态由业务逻辑设置。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/article"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建文章请求体。需要先准备 category_id 和 tag_ids，示例会创建一篇草稿文章",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"id\\\":1},\\\"msg\\\":\\\"创建文章成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数格式错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"创建文章失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/admin/articles/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员按文章 ID 删除文章。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/article"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"删除文章成果\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"文章ID不合法\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"删除文章失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/admin/comments/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员按评论 ID 删除评论。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin/comment"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "评论 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"msg\\\":\\\"删除评论成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"删除评论失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/api/cmments/get": {
            "get": {
                "description": "根据评论目标类型和目标 ID 分页获取评论列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "获取评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标类型，1 文章，2 留言板",
                        "name": "target_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目标 ID",
                        "name": "target_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"list\\\":[],\\\"total\\\":0,\\\"page\\\":1,\\\"page_size\\\":10},\\\"msg\\\":\\\"获取评论成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"获取评论失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/comments/add": {
            "post": {
                "description": "添加文章评论或留言板评论。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "添加评论",
                "parameters": [
                    {
                        "description": "评论请求体。文章评论 target_type 示例：1；留言 target_type 示例：2，留言示例可参考 CreateMessageRequest 字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "示例：{\\\"code\\\":200,\\\"data\\\":{\\\"id\\\":1},\\\"msg\\\":\\\"评论成功\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "示例：{\\\"code\\\":400,\\\"msg\\\":\\\"参数错误\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "示例：{\\\"code\\\":500,\\\"msg\\\":\\\"添加评论失败\\\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateArticleRequest": {
            "type": "object",
            "required": [
                "category_id",
                "content",
                "cover_url",
                "is_top",
                "summary",
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "content": {
                    "type": "string",
                    "example": "这里是文章正文内容，包含 Gin、GORM 和 MySQL 的实践记录。"
                },
                "cover_url": {
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "is_top": {
                    "type": "integer",
                    "example": 0
                },
                "summary": {
                    "type": "string",
                    "example": "这是一篇关于 Go 后端开发的文章摘要"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "我的第一篇博客"
                }
            }
        },
...
        "controllers.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "nickname",
                "target_type"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "这是一条 Apifox 测试评论"
                },
                "email": {
                    "type": "string",
                    "example": "tester@example.com"
                },
                "nickname": {
                    "type": "string",
                    "example": "Apifox Tester"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "target_id": {
                    "type": "integer",
                    "example": 1
                },
                "target_type": {
                    "type": "integer",
                    "example": 1
                },
                "website": {
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}